requirements.txt
amqp==2.4.0
appdirs==1.4.3
astroid==2.1.0
beautifulsoup4==4.7.0
billiard==3.5.0.5
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coverage==4.5.1
cssselect==1.0.3
dj-database-url==0.5.0
Django==2.1.2
django-compat==1.0.15
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-nose==1.4.6
django-templated-mail==1.1.1
djangorestframework==3.9.0
drf-nested-routers==0.91
fake-useragent==0.1.11
gunicorn==19.9.0
idna==2.8
isort==4.3.4
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
mccabe==0.6.1
nose==1.3.7
parse==1.9.0
psycopg2==2.7.6.1
pyee==5.0.0
pylint==2.2.2
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2018.7
six==1.12.0
soupsieve==1.6.1
tqdm==4.28.1
urllib3==1.24.1
vine==1.2.0
w3lib==1.19.0
websockets==6.0
whitenoise==4.1.2
wrapt==1.10.11
